Use attribute (rather than child) for property value
Prior to this commit, when writing out the properties
for a service component descriptor xml, the property
values were written as an XML child element inside
of the `property` tag.

In some OSGi implementations (Felix, at least), this
causes the service property to be initialized as
an array of String objects, rather than a single
instance of String.  Many existing OSGi services
that interrogate these properties seem to be
written to expect a String instance, rather than
an array of Strings; thus, these services fail
or cannot determine the values of the properties.

This commit simply changes the XML writer to
write the property values as a `value` attribute
on the `property` tag, rather than as a nested
child element.  (This is the behavior that the
felix-scr plugin provides.)  With this change,
services that were previously failing seem to
succeed.